# perf: implement Conv1D with im2col and GEMM