@minhthuc2502 2024-01-26
Thank you for your PR. Do you think with the larger model, the increasing in memory will be very important?  @nguyendc-systran  @vince62s What do you think ?