@ebraraktas 2024-01-27
You can find the allocation sizes for `large-v2`:

|   batch_size |   in_channel |   in_length |   out_channel |   out_length |   old_usage (KB) |   new_usage (KB) | relative_impact   |
|-------------:|-------------:|------------:|--------------:|-------------:|-----------------:|-----------------:|:------------------|
|            1 |           80 |        3000 |            80 |         3000 |              534 |              703 | +169 KB (+32%)    |
|            1 |         1280 |        3000 |          1280 |         1500 |             8550 |             5625 | -2925 KB (-34%)   |
|            2 |           80 |        3000 |            80 |         3000 |              768 |             1406 | +638 KB (+83%)    |
|            2 |         1280 |        3000 |          1280 |         1500 |            12300 |            11250 | -1050 KB (-9%)    |
|            4 |           80 |        3000 |            80 |         3000 |             1237 |             2812 | +1575 KB (+127%)  |
|            4 |         1280 |        3000 |          1280 |         1500 |            19800 |            22500 | +2700 KB (+14%)   |

Note that these are temporary tensors, and I don't think extra ~3MB for `batch_size=4` is a big amount.