@minhthuc2502 2024-02-09
No extra temporary memory is required. It looks good to me. I will merge this.